GOVSI-1100 - Validate new password against existing password
- When a user tries to either reset or change their password using their existing password then we should not allow that and return an error to the frontend. The frontend will look for this error and display an appropriate error message to the user.

+++ frontend-api/src/main/java/uk/gov/di/authentication/frontendapi/lambda/ResetPasswordHandler.java
@@ -97,11 +97,17 @@ public class ResetPasswordHandler
                                     return generateApiGatewayProxyErrorResponse(
                                             400, ErrorResponse.ERROR_1021);
                                 }
-                                codeStorageService.deleteSubjectWithPasswordResetCode(
-                                        resetPasswordWithCodeRequest.getCode());
                                 UserCredentials userCredentials =
                                         authenticationService.getUserCredentialsFromSubject(
                                                 subject.get());
+                                if (userCredentials
+                                        .getPassword()
+                                        .equals(resetPasswordWithCodeRequest.getPassword())) {
+                                    return generateApiGatewayProxyErrorResponse(
+                                            400, ErrorResponse.ERROR_1024);
+                                }
+                                codeStorageService.deleteSubjectWithPasswordResetCode(
+                                        resetPasswordWithCodeRequest.getCode());
                                 authenticationService.updatePassword(
                                         userCredentials.getEmail(),
                                         resetPasswordWithCodeRequest.getPassword());
